chore: fix timeout setup
I don't care if a test takes a while as long as the test is
doing something. I do care if a test is doing nothing for a
long time. This commit adjusts the timeout setup to be more
forgiving of slow tests and more strict on tests that actually
have a problem.

+ test timeout extended to 10 minutes
+ action timeout limited to 30 seconds
  + an action should not take longer than 30 seconds