fix: deal with infinite scroll when cleaning up
On some entity listing pages we do not refresh
the list after an entity is deleted. This means
we can remove all the entities on that page, but
not really have cleaned everything up since
we only load 50 entities at a time. In these cases
we need to scroll and load all the entities before
we can clean up.