fix: don't rely on dom to determine current filter
the email history page stores the state of the
current filters in session storage. this state
is applied to the DOM when the page is loaded
which is too slow for the tests and the controls
for the filters are not disable until they are set.

this means when applying a filter we need to
rely on the state in session storage to
determine if the filter we are trying to apply
is already applied.